requirements.txt
httpx
pandas
numpy
python-dotenv
python-telegram-bot
APScheduler
sqlalchemy
aiosqlite
web3